client side streaming seems to be working!
