+++ api/src/servers/utils/mod.rs
@@ -1,0 +1,1 @@
+pub mod sensor_stream;
