+++ drivers/Cargo.toml
@@ -7,13 +7,15 @@ edition = "2018"
 # See more keys and their definiticaons at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+proto = { path = "../proto"}
 serialport = "4.0.0"
 byteorder = "1.4.2"
 hex = "0.4.2"
 parse_int = "0.5.0"
 colored = "2.0.0"
-async-stream = "0.3.0"
 futures-core = "0.3.12"
 futures-util = "0.3.12"
 tokio = { version = "1.0", features = ["rt-multi-thread", "macros", "sync", "time"] }
+tokio-stream = { version =  "0.1", features = ["net"] }
+async-stream = "0.3.0"
 async-std = "1.9.0"
